requirements.txt
# Core dependencies
numpy>=1.24.0
opencv-python>=4.8.0
pymodbus>=3.6.0
pyserial>=3.5

# Computer Vision & AI
torch>=2.0.0
torchvision>=0.15.0
onnx>=1.15.0
onnxruntime-gpu>=1.16.0

# Jetson specific (install via JetPack)
# - pycuda
# - tensorrt
# - jetson-inference (optional)

# Navigation & Control
scipy>=1.11.0
scikit-learn>=1.3.0

# Utilities
pyyaml>=6.0
loguru>=0.7.0
python-dotenv>=1.0.0

# Visualization & Debugging
matplotlib>=3.7.0
pillow>=10.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional monitoring
# jetson-stats (install separately: sudo pip3 install jetson-stats)
